Basic Collectables Implemented
I hope this is the right format.
Basic collectables that rotate and disappear when intersected with
have been implemented. This includes their material, as well as changes
to layers and tags. A new "Collectable" layer has been added that does not
collide with the tank "feet" and any part of the tank body, which the
collectables have been assigned to. The tank's cube and nose have been
added to the "Player" layer for this reason as well. The Collectables also
have the tag "Collectable" but this can probably be removed. Lastly,
unless the code can changed, only the "TankBody_1" object should have the
tag "Player", as the CollectableController script relies on this
to find the Player GameObject. The CollectableController script rotates
the Collectables and checks for intersection with the Player bounds.